gulf jobs
M Gheewala Recruitment Agency for Gulf Jobs
The gulf and the middle-east jobs, sectors have been active ever since the 70s. The construction, engineering, and infrastructure were beginning to boom and take a big shape in the scheme of things for organizations and the economy there. But where then would you get the right people, manpower, staff, skilled, semi-skilled workers? Nearly 40 years when the demand for this kind of personnel was extremely high and shooting…who would fulfill their requirements? It was not going to be an easy task. Getting the brief from the clients and studying them thoroughly was important as the Indian manpower resources although came from good stock, yet there were fears of working in countries with different laws and cultures.

The fears were from sides. But M Gheewala HR Solutions came as one big troubleshooter and relief for meeting the requirements of both the interested parties – the employer on one side and the employee on the other. M Gheewala HR Solutions actually bridged the yawning gap and became a force to reckon with very soon. It was a huge blessing for all.

A detailed study by our young professionals even in those days made us a brand entity as "people who could arrange the best people for all kinds of job openings in the gulf." A reputation that precedes us to date. And this reputation has been further solidified and fortified over the years. Today with a serving period of over 40 years job seekers recollect M Gheewala HR Solutions with awe and fondness.
https://www.mgheewala.com/gulf-jobs.php